requirements.txt
accelerate==0.28.0
datasets==2.18.0
huggingface_hub==0.21.4
numpy==1.26.4
open_clip_torch==2.24.0
pandas==2.2.1
peft==0.9.0
Pillow==10.3.0
timm==0.9.8
torch==2.2.1
torch==2.2.0
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.1
trl==0.8.2.dev0
